+++ java/shared/java/org/signal/libsignal/internal/NativeTesting.kt
@@ -316,6 +316,8 @@ public object NativeTesting {
   @JvmStatic
   public external fun TESTING_SetRegistrationLockTests(): Array<Object>
   @JvmStatic
+  public external fun TESTING_SetRegistrationRecoveryPasswordTests(): Array<Object>
+  @JvmStatic
   public external fun TESTING_SetUsernameLinkTests(): Array<Object>
   @JvmStatic
   public external fun TESTING_SignedPublicPreKey_CheckBridgesCorrectly(sourcePublicKey: ObjectHandle, signedPreKey: SignedPublicPreKey<*>): Unit